Project 5c: Do-While loop
This project will prompt the user for the input of an employee's
id number, first name, last name, department name, job title and salary
level.
The users inputs would be concatenated into a string along with a new
line
character and appended to a data file. Once the user enters the word
"quit"
the program will terminate.